feedback on typed_dicts.py
