+++ scripts/cat/factories/typed_dicts.py
@@ -1,5 +1,7 @@
 from typing import TypedDict, Optional, List, Dict
 
+from scripts.cat.enums import CatSocial, CatRank, CatAge
+
 
 class GenderDict(TypedDict, total=False):
     sex: str
@@ -23,11 +25,6 @@ class CatTogglesDict(TypedDict):
     favourite: bool
 
 
-class MateshipDict(TypedDict):
-    mate: List[str]
-    previous_mates: List[str]
-
-
 class InheritanceDict(TypedDict):
     parent1: Optional[str]
     parent2: Optional[str]
@@ -40,3 +37,25 @@ class InheritanceDict(TypedDict):
 class AfterlifeAffinityDict(TypedDict):
     starclan: int
     dark_forest: int
+
+
+class StatusDict(TypedDict, total=False):
+    """
+    Dict containing:
+
+    "group_history": list[dict],
+    "standing_history": list[dict],
+    "social": CatSocial,
+    "group": CatGroup
+    "rank": CatRank
+    "age": CatAge
+
+    Dict does not need to contain all keys. However, if you have no group history, then you must include a rank or age
+    """
+
+    group_history: Optional[List[Dict]]
+    standing_history: Optional[List[Dict]]
+    social: Optional[CatSocial]
+    group_ID: Optional[str]
+    rank: Optional[CatRank]
+    age: Optional[CatAge]
